+++ tests/different-conditions-fold/main_ftr.c
@@ -1,0 +1,11 @@
+     return 1;
+  } else {
+     return 0;
+  }
+}
+
+int main(int argc, char* argv[]) {
+  foo(0, 1);
+
+  return 0;
+}