+++ tests/different-conditions-strict/main1.c
@@ -1,0 +1,1 @@
+  if (param != 0 && param2 != 1) {